move kernels into collisions dir
