+++ PySDM/dynamics/__init__.py
@@ -3,10 +3,10 @@ Classes representing physicochemical processes:
 `PySDM.dynamics.coalescence.Coalescence`,
 `PySDM.dynamics.condensation.Condensation`, ...
 """
-from .coalescence import Coalescence
-from .condensation import Condensation
-from .displacement import Displacement
-from .eulerian_advection import EulerianAdvection
-from .ambient_thermodynamics import AmbientThermodynamics
-from .aqueous_chemistry import AqueousChemistry
-from .freezing import Freezing
+from PySDM.dynamics.collisions.coalescence import Coalescence
+from PySDM.dynamics.condensation import Condensation
+from PySDM.dynamics.displacement import Displacement
+from PySDM.dynamics.eulerian_advection import EulerianAdvection
+from PySDM.dynamics.ambient_thermodynamics import AmbientThermodynamics
+from PySDM.dynamics.aqueous_chemistry import AqueousChemistry
+from PySDM.dynamics.freezing import Freezing
