--- PySDM/dynamics/coalescence_kernels/__init__.py
@@ -1,10 +1,0 @@
-"""
-Coalescence kernels including
-[Golovin](https://atmos-cloud-sim-uj.github.io/PySDM/physics/coalescence_kernels/golovin.html),
-[Geometric](https://atmos-cloud-sim-uj.github.io/PySDM/physics/coalescence_kernels/geometric.html)
-and other...
-"""
-from .golovin import Golovin
-from .geometric import Geometric
-from .hydrodynamic import Hydrodynamic
-from .electric import Electric
