+++ PySDM/dynamics/collisions/__init__.py
@@ -1,0 +1,1 @@
+from .coalescence import Coalescence
