+++ PySDM/dynamics/collisions/kernels/__init__.py
@@ -1,0 +1,10 @@
+"""
+Coalescence kernels including
+[Golovin](https://atmos-cloud-sim-uj.github.io/PySDM/physics/collisions/kernels/golovin.html),
+[Geometric](https://atmos-cloud-sim-uj.github.io/PySDM/physics/collisions/kernels/geometric.html)
+and other...
+"""
+from .geometric import Geometric
+from  .golovin import Golovin
+from .electric import Electric
+from .hydrodynamic import Hydrodynamic
