+++ PySDM/dynamics/collisions/kernels/electric.py
@@ -2,7 +2,7 @@
 kernel modelling influence of electric field of 3000V/cm
  as in [Berry 1967](https://doi.org/10.1175/1520-0469(1967)024%3C0688:CDGBC%3E2.0.CO;2)
 """
-from PySDM.dynamics.coalescence_kernels.impl.parameterized import Parameterized
+from PySDM.dynamics.collisions.kernels.impl.parameterized import Parameterized
 
 
 class Electric(Parameterized):
